+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Next
 
 * PR #69: Fixed dependency shading, reduced the distributable jar size
+* PR #76: Added support for service account impersonation
 
 ## 2.0.0-preview - 2023-04-12